refactor: move the package to a more appropriate location
