+++ product/src/main/java/cm/twentysix/product/domain/model/Product.java
@@ -2,7 +2,7 @@ package cm.twentysix.product.domain.model;
 
 import cm.twentysix.product.dto.CreateProductForm;
 import cm.twentysix.product.dto.UpdateProductForm;
-import cm.twentysix.product.service.dto.CategoryInfoDto;
+import cm.twentysix.product.dto.CategoryInfoDto;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
