+++ product/src/main/java/cm/twentysix/product/service/CategoryService.java
@@ -1,7 +1,7 @@
 package cm.twentysix.product.service;
 
 import cm.twentysix.product.domain.repository.CategoryAggregationRepository;
-import cm.twentysix.product.service.dto.CategoryInfoDto;
+import cm.twentysix.product.dto.CategoryInfoDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
